Move shapes out of canvas
